Create a concern for user authentication helpers
